+++ api/v1alpha1/monocle_types.go
@@ -22,7 +22,9 @@ import (
 
 // MonocleSpec defines the desired state of Monocle
 type MonocleSpec struct {
-	FQDN string `json:"FQDN"`
+	FQDN             string `json:"FQDN"`
+	StorageClassName string `json:"storageClassName,omitempty"`
+	StorageSize      string `json:"storageSize,omitempty"`
 }
 
 // MonocleStatus defines the observed state of Monocle